♻️ Feat(#79): 셀 뷰 관리 역할 집중 개선
