+++ ios/issue-tracker-01/issue-tracker-01/Issue/IssueEditor/EditorOptionCell.swift
@@ -11,11 +11,7 @@ class EditorOptionCell: UITableViewCell {
     
     static let identifier: String = "EditorOptionCell"
     
-    private var options: [OptionType]? = [] {
-        didSet {
-            self.collectionView.reloadData()
-        }
-    }
+    private var dataSource = EditorOptionDataSource()
 
     @IBOutlet weak var titleLabel: UILabel!
     @IBOutlet weak var collectionView: UICollectionView!
@@ -37,8 +33,8 @@ class EditorOptionCell: UITableViewCell {
             UINib(nibName: "EditorOptionLabelCell", bundle: .main),
             forCellWithReuseIdentifier: EditorOptionLabelCell.identifier
         )
-        self.collectionView.dataSource = self
-        self.collectionView.delegate = self
+        self.collectionView.dataSource = dataSource
+        self.collectionView.delegate = dataSource
     }
 
     override func setSelected(_ selected: Bool, animated: Bool) {
@@ -61,42 +57,7 @@ class EditorOptionCell: UITableViewCell {
     }
     
     func configureOptions(options: [OptionType]?) {
-        self.options = options
+        self.dataSource.options = options
         self.collectionView.reloadData()
-    }
-}
-
-extension EditorOptionCell: UICollectionViewDataSource, UICollectionViewDelegate {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return options?.count ?? 0
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: EditorOptionLabelCell.identifier, for: indexPath) as? EditorOptionLabelCell else {
-            return UICollectionViewCell()
-        }
-        
-        if let option = options?[indexPath.item] {
-            
-            cell.setLabel(with: option)
-        }
-        
-        return cell
-    }
-}
-
-extension EditorOptionCell: UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        guard let option = options?[indexPath.item] else {
-            return CGSize(width: 0, height: 0)
-        }
-        
-        let label = UILabel()
-        let padding: CGFloat = 10
-        label.applyStyle(fontManager: FontManager(weight: .bold, size: .medium), textColor: .gray50)
-        label.text = option.displayText
-        
-        let width = label.intrinsicContentSize.width + padding
-        return CGSize(width: width, height: 24)
     }
 }
